Heroku tweaks.
Trimmed some unnecessary files from the Heroku slug.
Changed :heroku-mode to detect AWS credentials in the environment rather
than LEIN_NO_DEV, which was a hack for Leiningen 1.x so I probably
shouldn't rely on it. AWS credentials need to be there anyway and are a
better indicator of the runtime environment.